+++ src/modules/kiosks/components/Toolbar.js
@@ -1,14 +1,27 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
-import { Segment, Grid, Input, Button, Icon, Dropdown, Divider} from 'semantic-ui-react';
+import { connect } from 'react-redux';
+import {
+  Segment,
+  Grid,
+  Input,
+  Button,
+  Icon,
+  Dropdown,
+  Divider,
+} from 'semantic-ui-react';
 
-const Toolbar = ({ search,
+import { getActiveUserIDState } from '../../users/selectors';
+
+const Toolbar = ({
+  search,
   setSearch,
   kiosks,
   setKiosk,
   kiosksStatus,
   setKioskStatus,
+  rootUser,
   // kiosksNetworkStatus, //!LIV-2285
   // setKioskNetworkStatus //!LIV-2285
 }) => {
@@ -17,11 +30,11 @@ const Toolbar = ({ search,
   };
   const handleKioskChange = (e, { value }) => {
     const text = value === 'All' ? '' : value;
-    setKiosk(text)
+    setKiosk(text);
   };
   const handleKiosksStatus = (e, { value }) => {
     const text = value === 'All' ? '' : value;
-    setKioskStatus(text)
+    setKioskStatus(text);
   };
   // const handleKiosksNetworkStatus = (e, { value }) => { //!LIV-2285
   //   const text = value === 'All' ? '' : value;
@@ -40,47 +53,47 @@ const Toolbar = ({ search,
               className="full-width"
             />
           </Grid.Column>
-
-          <Grid.Column textAlign="right">
-            <Button
-              icon
-              labelPosition="left"
-              color="green"
-              compact
-              as={Link}
-              // disabled={true} //! Disabled temporarily #golive2
-              to="/kiosks/edit/new" 
-            >
-              <Icon name="right arrow" />
-              Add Kiosk
-            </Button>
-          </Grid.Column>
+          {rootUser && (
+            <Grid.Column textAlign="right">
+              <Button
+                icon
+                labelPosition="left"
+                color="green"
+                compact
+                as={Link}
+                to="/kiosks/edit/new"
+              >
+                <Icon name="right arrow" />
+                Add Kiosk
+              </Button>
+            </Grid.Column>
+          )}
         </Grid.Row>
-      
-      <Divider style={{marginTop: 0, marginBottom: 0}} />
 
-      <Grid.Row verticalAlign="middle" columns="equal">
-        <Grid.Column width={3}>
-          <Dropdown
-              placeholder="All Kiosks"
+        <Divider style={{ marginTop: 0, marginBottom: 0 }} />
+
+        <Grid.Row verticalAlign="middle" columns="equal">
+          <Grid.Column width={3}>
+            <Dropdown
+              placeholder="Kiosks"
               selection
               className="full-width"
               onChange={handleKioskChange}
               options={kiosks}
-          />
-        </Grid.Column>
+            />
+          </Grid.Column>
 
-        <Grid.Column width={3}>
-          <Dropdown
+          <Grid.Column width={3}>
+            <Dropdown
               placeholder="Door Status"
               selection
               className="full-width"
               onChange={handleKiosksStatus}
               options={kiosksStatus}
-          />
-        </Grid.Column>
+            />
+          </Grid.Column>
 
-        {/* <Grid.Column width={3}> //!LIV-2285
+          {/* <Grid.Column width={3}> //!LIV-2285
           <Dropdown
               placeholder="Network Status"
               selection
@@ -89,8 +102,7 @@ const Toolbar = ({ search,
               options={kiosksNetworkStatus}
           />
         </Grid.Column> */}
-      </Grid.Row>
-      
+        </Grid.Row>
       </Grid>
     </Segment>
   );
@@ -101,4 +113,8 @@ Toolbar.propTypes = {
   setSearch: PropTypes.func.isRequired,
 };
 
-export default Toolbar;
+const mapStateToProps = state => ({
+  rootUser: state.user.root,
+});
+
+export default connect(mapStateToProps)(Toolbar);